+++ src/__tests__/cli-capture.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { runCli } from '../cli.ts';
-import type { DaemonRequest, DaemonResponse } from '../daemon-client.ts';
+import type { DaemonRequest, DaemonResponse } from '../daemon/client/daemon-client.ts';
 import { installIsolatedCliTestEnv } from './cli-test-env.ts';
 
 class ExitSignal extends Error {
