+++ src/__tests__/cli-react-devtools-session.test.ts
@@ -15,7 +15,7 @@ import {
   hashRemoteConfigFile,
   writeRemoteConnectionState,
 } from '../remote/remote-connection-state.ts';
-import type { DaemonResponse } from '../daemon-client.ts';
+import type { DaemonResponse } from '../daemon/client/daemon-client.ts';
 
 afterEach(() => {
   vi.clearAllMocks();
