+++ src/__tests__/daemon-client-progress.test.ts
@@ -3,7 +3,7 @@ import { EventEmitter } from 'node:events';
 import type { Socket } from 'node:net';
 import { test } from 'vitest';
 import type { DaemonRequest, DaemonResponse } from '../daemon/types.ts';
-import { readDaemonSocketProgressResponse } from '../daemon-client-progress.ts';
+import { readDaemonSocketProgressResponse } from '../daemon/client/daemon-client-progress.ts';
 import { AppError } from '../kernel/errors.ts';
 
 type MockSocket = EventEmitter & {
