+++ src/cli/commands/proxy.ts
@@ -1,7 +1,10 @@
 import { randomBytes } from 'node:crypto';
 import { createDaemonProxyServer } from '../../remote/daemon-proxy.ts';
 import { buildDaemonHttpBaseUrl } from '../../daemon/http-contract.ts';
-import { ensureDaemon, resolveClientSettings } from '../../daemon-client-lifecycle.ts';
+import {
+  ensureDaemon,
+  resolveClientSettings,
+} from '../../daemon/client/daemon-client-lifecycle.ts';
 import { AppError } from '../../kernel/errors.ts';
 import { colorize, supportsColor } from '../../utils/output.ts';
 import type { CliFlags } from '../parser/cli-flags.ts';
